requirements.txt
browserstack-local >= 1.2.3
selenium
browserstack-sdk
pytest
python-dotenv
